2nd revision

+++ python/Lab10_unitconversion.py
@@ -6,7 +6,7 @@ Marcel Schaeffer
 
 #user types amount of feet
 
-feet = int(input('Type an amount of feet: '))
+feet = float(input('Type an amount of feet: '))
 unit = input('Pick units to convert feet into; i.e. meters, miles, kilometers: ')
 
 #units in meters
